Anti-piracy protection: automatically disable and remove blacklisted piracy apps
registers a broadcast receiver that detects blacklisted package installs and also
activates on ACTION_BOOT_COMPLETED to make sure they can't be sideloaded or flashed in
recovery.

If a package is detected a service starts that automatically disables and uninstalls the
blacklisted piracy app then stops itself again using reflection of an internal method
in package manager to do it the same way any other app is uninstalled.

Change-Id: Ic4d52d48edb1d9f99f1f782f8f6dd4766ae10a3d